pyproject.toml
[project]
name = "orders-service"
version = "0.1.0"
description = "Test assignment: Orders Service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aio-pika>=9.4.0",
  "alembic>=1.13.1",
  "asyncpg>=0.29.0",
  "celery-types>=0.24.0",
  "celery[redis]>=5.3.6",
  "email-validator>=2.1.0",
  "fastapi>=0.110.0",
  "orjson>=3.10.0",
  "passlib[bcrypt]>=1.7.4",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.1",
  "python-dotenv>=1.0.0",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.9",
  "redis>=5.0.1",
  "slowapi>=0.1.9",
  "sqlalchemy[asyncio]>=2.0.25",
  "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "mypy>=1.8.0",
  "pytest>=8.0.0",
  "ruff>=0.3.0",
  "types-passlib>=1.7.7.13",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
